# protocol/openapi.yaml
openapi: 3.1.1
info:
  title: VaultDatum API
  version: 0.1.0
  description: |
    The protocol contract between the authoritative VaultDatum server and its
    replica clients. Realtime notifications may trigger synchronization, but
    correctness depends only on these HTTP operations.

    A private-network deployment relies on its protected network. In a
    public-token deployment, every /api/v1 HTTP operation requires the
    VaultBearerToken Authorization header; notification WebSocket upgrades use
    a one-time RealtimeTicket instead. Tokens and tickets never appear in URLs.
servers:
  - url: http://localhost:8080
...
paths:
...
  /api/v1/content:
    get:
      tags:
        - Synchronization
      operationId: downloadContent
      summary: Download content only while its path state matches
      description: |
        `revision` is the latest revision of this path, not the current global
        Vault revision. If the path changes before the content is read, the
        server returns `STATE_CHANGED` instead of historical content.
      parameters:
        - $ref: "#/components/parameters/SyncPathParameter"
        - $ref: "#/components/parameters/PathRevisionParameter"
        - $ref: "#/components/parameters/ContentHashParameter"
      responses:
        "200":
          description: Current file content
          headers:
            X-VaultDatum-Content-Hash:
              description: Hash of the returned bytes
              schema:
                $ref: "#/components/schemas/ContentHash"
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        "400":
          $ref: "#/components/responses/InvalidRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          description: The current path state no longer matches the request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProtocolError"
        "503":
          $ref: "#/components/responses/ServerUnavailable"
  /api/v1/history:
    get:
      tags:
        - History
      operationId: listFileHistory
      summary: List the committed changes of one file, newest first
      description: |
        Used only to restore a single file. Synchronization never depends on
        history. Entries whose content is no longer kept report
        `contentAvailable: false`.
      parameters:
        - $ref: "#/components/parameters/SyncPathParameter"
        - name: before
          in: query
          required: false
          description: Return entries with a revision strictly lower than this one.
          schema:
            $ref: "#/components/schemas/CommittedRevision"
        - name: limit
          in: query
          required: false
          description: Maximum number of entries to return.
          schema:
            type: integer
            format: int32
            minimum: 1
            maximum: 100
            default: 50
      responses:
        "200":
          description: File history page
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HistoryPage"
        "400":
          $ref: "#/components/responses/InvalidRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "503":
          $ref: "#/components/responses/ServerUnavailable"
  /api/v1/history/content:
    get:
      tags:
        - History
      operationId: downloadHistoryContent
      summary: Download kept content by hash
      description: |
        Returns content kept in history or held by a current Vault file. The
        server verifies the hash before returning bytes.
      parameters:
        - name: contentHash
          in: query
          required: true
          description: Hash of the requested content.
          schema:
            $ref: "#/components/schemas/ContentHash"
      responses:
        "200":
          description: Kept content
          headers:
            X-VaultDatum-Content-Hash:
              description: Hash of the returned bytes
              schema:
                $ref: "#/components/schemas/ContentHash"
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        "400":
          $ref: "#/components/responses/InvalidRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          description: The content is no longer kept
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProtocolError"
        "503":
          $ref: "#/components/responses/ServerUnavailable"
...
components:
  parameters:
...
    SyncPathParameter:
      name: path
      in: query
      required: true
      description: Vault-root-relative logical path using `/` separators.
      schema:
        $ref: "#/components/schemas/SyncPath"
    PathRevisionParameter:
      name: revision
      in: query
      required: true
      description: Latest committed revision of the requested path.
      schema:
        $ref: "#/components/schemas/CommittedRevision"
    ContentHashParameter:
      name: hash
      in: query
      required: true
      description: Expected hash of the requested file content.
      schema:
        $ref: "#/components/schemas/ContentHash"
  responses:
    InvalidRequest:
      description: Request syntax or input is invalid
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ProtocolError"
    Unauthorized:
      description: Authentication is required
      headers:
        WWW-Authenticate:
          description: Bearer challenge for a public-token deployment.
          schema:
            type: string
            const: Bearer realm="vaultdatum"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ProtocolError"
    Forbidden:
      description: The caller is not allowed to access this Vault
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ProtocolError"
    ServerUnavailable:
      description: The server is recovering or is not ready
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ProtocolError"
  securitySchemes:
    VaultBearerToken:
      type: http
      scheme: bearer
      bearerFormat: VaultDatum Vault token
  schemas:
...
    ClientId:
      type: string
      minLength: 1
      description: Stable identity of one client installation.
...
    CommittedRevision:
      type: integer
      format: int64
      minimum: 1
      description: Server-assigned global commit order.
    SyncPath:
      type: string
      minLength: 1
      description: |
        Vault-root-relative logical path. Absolute paths, traversal segments,
        backslashes, and `.obsidian` are invalid.
    ContentHash:
      type: string
      pattern: "^sha256:[0-9a-f]{64}$"
      description: Canonical SHA-256 hash (`sha256:` plus lowercase hexadecimal).
...
    ChangeType:
      type: string
      enum:
        - CREATE
        - MODIFY
        - DELETE
        - RENAME
        - MOVE
...
    HistoryPage:
      type: object
      additionalProperties: false
      required:
        - path
        - entries
        - hasMore
      properties:
        path:
          $ref: "#/components/schemas/SyncPath"
        entries:
          type: array
          items:
            $ref: "#/components/schemas/HistoryEntry"
        hasMore:
          type: boolean
    HistoryEntry:
      type: object
      additionalProperties: false
      required:
        - revision
        - type
        - committedAt
        - actor
        - state
        - contentAvailable
      properties:
        revision:
          $ref: "#/components/schemas/CommittedRevision"
        type:
          $ref: "#/components/schemas/ChangeType"
        committedAt:
          type: string
          format: date-time
        actor:
          $ref: "#/components/schemas/ChangeActor"
        state:
          type: string
          enum:
            - PRESENT
            - DELETED
        contentHash:
          $ref: "#/components/schemas/ContentHash"
        size:
          type: integer
          format: int64
          minimum: 0
        contentAvailable:
          type: boolean
          description: True when the content of this entry can be downloaded.
        previousPath:
          $ref: "#/components/schemas/SyncPath"
...
    ChangeActor:
      oneOf:
        - $ref: "#/components/schemas/ClientChangeActor"
        - $ref: "#/components/schemas/ServerChangeActor"
    ClientChangeActor:
      type: object
      additionalProperties: false
      required:
        - type
        - clientId
      properties:
        type:
          type: string
          const: CLIENT
        clientId:
          $ref: "#/components/schemas/ClientId"
    ServerChangeActor:
      type: object
      additionalProperties: false
      description: |
        Server-originated change. `SERVER_EXTERNAL` is recorded only for the
        one-time initial import of files copied into a new server Vault;
        direct server Vault edits are never imported as changes. `SYSTEM` is
        reserved.
      required:
        - type
      properties:
        type:
          type: string
          enum:
            - SERVER_EXTERNAL
            - SYSTEM
...
    ProtocolError:
      type: object
      additionalProperties: false
      required:
        - error
      properties:
        error:
          $ref: "#/components/schemas/ErrorDetail"
    ErrorDetail:
      type: object
      additionalProperties: false
      required:
        - code
      properties:
        code:
          type: string
          enum:
            - INVALID_REQUEST
            - INVALID_PATH
            - PATH_EXCLUDED
            - VAULT_MISMATCH
            - BASE_STATE_MISMATCH
            - OPERATION_ID_REUSED
            - CONTENT_HASH_MISMATCH
            - STATE_CHANGED
            - HISTORY_NOT_AVAILABLE
            - MANIFEST_EXPIRED
            - MANIFEST_NOT_FOUND
            - RECOVERY_REQUIRED
            - CONTENT_NOT_RETAINED
            - SERVER_NOT_READY
            - UNAUTHORIZED
            - FORBIDDEN
        message:
          type: string
        details:
          type: object
          additionalProperties: true